Add useUpdates hook
