+++ src/components/SagaMiddlewareProvider.js
@@ -1,6 +1,5 @@
 import React, { createContext, useContext, useEffect } from 'react';
 
-
 const SagaMiddlewareContext = createContext();
 
 export const useSaga = (saga, deps) => {
